requirements.txt
tensorflow-gpu==1.14.0
opencv-python
opencv-contrib-python
numpy==1.16.0
